+++ shared/actions.py
@@ -2213,6 +2213,17 @@ async def change_virtdisk_enable(enable):
         glob.VD.shutdown()
         assert not glob.VD
 
+async def change_seed_vault(is_enabled):
+    # user has changed seed vault enable/disable flag
+    from glob import settings
+    if not is_enabled and settings.get('seeds'):
+        # restore it
+        settings.set('seedvault', True)
+        ch = await ux_show_story("Please remove all seeds from the vault before disabling")
+        return
+
+    goto_top_menu()
+
 async def change_which_chain(*a):
     # setting already changed, but reflect that value in other settings
     try: