+++ shared/pwsave.py
@@ -132,7 +132,7 @@ class PassphraseSaver:
         async def doit(menu, idx, item):
             # apply the password immediately and drop them at top menu
             pw, expect_xfp = item.arg
-            set_bip39_passphrase(pw)
+            await set_bip39_passphrase(pw)
 
             from glob import settings
             from utils import xfp2str