+++ testing/test_ephemeral.py
@@ -96,7 +96,7 @@ def goto_eph_seed_menu(goto_home, pick_menu_item, cap_story, need_keypress):
 def verify_ephemeral_secret_ui(cap_story, need_keypress, cap_menu, dev, goto_home, pick_menu_item,
                                fake_txn, try_sign, goto_eph_seed_menu, reset_seed_words,
                                get_identity_story, get_seed_value_ux):
-    def doit(mnemonic=None, xpub=None):
+    def doit(mnemonic=None, xpub=None, seed_vault=False):
         time.sleep(0.3)
         _, story = cap_story()
         in_effect_xfp = story[1:9]
@@ -105,6 +105,52 @@ def verify_ephemeral_secret_ui(cap_story, need_keypress, cap_menu, dev, goto_hom
 
         menu = cap_menu()
         assert menu[0] == "Ready To Sign"  # returned to main menu
+
+        if seed_vault:
+            # check seed is saved in cache
+            pick_menu_item("Seed Vault")
+            time.sleep(.1)
+            sc_menu = cap_menu()
+            assert len(sc_menu) == 1  # stored seed
+            for i in sc_menu:
+                if in_effect_xfp in i.split()[-1]:
+                    pick_menu_item(i)
+                    time.sleep(.1)
+                    m = cap_menu()
+                    assert "Use This Seed" in m
+                    assert "Delete" in m
+                    assert "Rename" in m
+                    assert len(m) == 4  # xfp is top item (works same as "Use this seed")
+                    # apply it - even if it is applied already
+                    pick_menu_item("Use This Seed")
+                    time.sleep(.1)
+                    _, story = cap_story()
+                    assert "Press (1)" not in story
+                    assert "key in effect until next power down." in story
+                    assert in_effect_xfp in story
+                    need_keypress("y")
+                    pick_menu_item("Seed Vault")
+                    pick_menu_item(i)
+                    time.sleep(.1)
+                    # delete it from cache
+                    pick_menu_item("Delete")
+                    time.sleep(.1)
+                    _, story = cap_story()
+                    assert "Delete" in story
+                    assert in_effect_xfp in story
+                    need_keypress("y")
+                    time.sleep(.1)
+                    m = cap_menu()
+                    assert "(none saved yet)" in m
+                    assert len(m) == 1
+                    break
+            else:
+                raise pytest.fail("Failed to cache seed?")
+        else:
+            # Seed Vault disabled
+            m = cap_menu()
+            assert "Seed Vault" not in m
+
 
         ident_story = get_identity_story()
         assert "Ephemeral seed is in effect" in ident_story
@@ -140,11 +186,15 @@ def verify_ephemeral_secret_ui(cap_story, need_keypress, cap_menu, dev, goto_hom
 
 @pytest.mark.parametrize("num_words", [12, 24])
 @pytest.mark.parametrize("dice", [False, True])
+@pytest.mark.parametrize("seed_vault", [False, True])
 def test_ephemeral_seed_generate(num_words, pick_menu_item, goto_home, cap_story, need_keypress,
                                  reset_seed_words, goto_eph_seed_menu, dice, ephemeral_seed_disabled,
-                                 verify_ephemeral_secret_ui):
-
+                                 verify_ephemeral_secret_ui, seed_vault, settings_set, settings_get):
     reset_seed_words()
+    if seed_vault:
+        settings_set("seedvault", True)
+    else:
+        settings_set("seedvault", None)
     try:
         goto_eph_seed_menu()
     except:
@@ -172,18 +222,26 @@ def test_ephemeral_seed_generate(num_words, pick_menu_item, goto_home, cap_story
 
     need_keypress("6")  # skip quiz
     need_keypress("y")  # yes - I'm sure
-    time.sleep(0.1)
-    need_keypress("4")  # understand consequences
-    verify_ephemeral_secret_ui(mnemonic=e_seed_words)
+
+    if seed_vault:
+        time.sleep(0.1)
+        _, story = cap_story()
+        assert "Press (1) to store ephemeral secret into Seed Vault" in story
+        need_keypress("1")  # store it
+        need_keypress("y")  # confirm saved to Seed Vault
+
+    verify_ephemeral_secret_ui(mnemonic=e_seed_words, seed_vault=seed_vault)
 
 
 @pytest.mark.parametrize("num_words", [12, 18, 24])
 @pytest.mark.parametrize("nfc", [False, True])
 @pytest.mark.parametrize("truncated", [False, True])
+@pytest.mark.parametrize("seed_vault", [False, True])
 def test_ephemeral_seed_import_words(nfc, truncated, num_words, cap_menu, pick_menu_item, goto_home,
                                      cap_story, need_keypress, reset_seed_words, goto_eph_seed_menu,
                                      word_menu_entry, nfc_write_text, verify_ephemeral_secret_ui,
-                                     ephemeral_seed_disabled, get_seed_value_ux):
+                                     ephemeral_seed_disabled, get_seed_value_ux, seed_vault,
+                                     settings_set):
     if truncated and not nfc: return
 
     wordlists = {
@@ -194,6 +252,12 @@ def test_ephemeral_seed_import_words(nfc, truncated, num_words, cap_menu, pick_m
     words, expect_xfp = wordlists[num_words]
 
     reset_seed_words()
+
+    if seed_vault:
+        settings_set("seedvault", True)
+    else:
+        settings_set("seedvault", None)
+
     try:
         goto_eph_seed_menu()
     except:
@@ -219,10 +283,16 @@ def test_ephemeral_seed_import_words(nfc, truncated, num_words, cap_menu, pick_m
             nfc_write_text(truncated_words)
         else:
             nfc_write_text(words)
+        time.sleep(.5)
 
-    need_keypress("4")  # understand consequences
+    if seed_vault:
+        time.sleep(.1)
+        _, story = cap_story()
+        assert "Press (1) to store ephemeral secret into Seed Vault" in story
+        need_keypress("1")  # store it
+        need_keypress("y")  # confirm saved to Seed Vault
 
-    verify_ephemeral_secret_ui(mnemonic=words.split(" "))
+    verify_ephemeral_secret_ui(mnemonic=words.split(" "), seed_vault=seed_vault)
 
     nfc_seed = get_seed_value_ux(nfc=True)  # export seed via NFC (always truncated)
     seed_words = get_seed_value_ux()
@@ -232,11 +302,17 @@ def test_ephemeral_seed_import_words(nfc, truncated, num_words, cap_menu, pick_m
 @pytest.mark.parametrize("way", ["sd", "vdisk", "nfc"])
 @pytest.mark.parametrize('retry', range(3))
 @pytest.mark.parametrize("testnet", [True, False])
+@pytest.mark.parametrize("seed_vault", [False, True])
 def test_ephemeral_seed_import_tapsigner(way, retry, testnet, pick_menu_item, cap_story, enter_hex,
                                          need_keypress, reset_seed_words, goto_eph_seed_menu,
                                          verify_ephemeral_secret_ui, ephemeral_seed_disabled,
-                                         nfc_write_text, tapsigner_encrypted_backup):
+                                         nfc_write_text, tapsigner_encrypted_backup, settings_set,
+                                         seed_vault):
     reset_seed_words()
+    if seed_vault:
+        settings_set("seedvault", True)
+    else:
+        settings_set("seedvault", None)
 
     fname, backup_key_hex, node = tapsigner_encrypted_backup(way, testnet=testnet)
 
@@ -281,15 +357,24 @@ def test_ephemeral_seed_import_tapsigner(way, retry, testnet, pick_menu_item, ca
     assert "back of the card" in story
     need_keypress("y")  # yes I have backup key
     enter_hex(backup_key_hex)
-    verify_ephemeral_secret_ui(xpub=node.hwif())
+
+    if seed_vault:
+        time.sleep(.1)
+        _, story = cap_story()
+        assert "Press (1) to store ephemeral secret into Seed Vault" in story
+        need_keypress("1")  # store it
+        need_keypress("y")  # confirm saved to Seed Vault
+
+    verify_ephemeral_secret_ui(xpub=node.hwif(), seed_vault=seed_vault)
 
 
 @pytest.mark.parametrize("fail", ["wrong_key", "key_len", "plaintext", "garbage"])
 def test_ephemeral_seed_import_tapsigner_fail(cap_menu, pick_menu_item, goto_home, cap_story, fail,
                                               need_keypress, reset_seed_words, enter_hex,
                                               tapsigner_encrypted_backup, goto_eph_seed_menu,
-                                              microsd_path, ephemeral_seed_disabled):
+                                              microsd_path, ephemeral_seed_disabled, settings_set):
     reset_seed_words()
+    settings_set("seedvault", None)
     fail_msg = "Decryption failed - wrong key?"
     fname, backup_key_hex, node = tapsigner_encrypted_backup("sd", testnet=False)
     if fail == "plaintext":
@@ -350,11 +435,12 @@ def test_ephemeral_seed_import_tapsigner_fail(cap_menu, pick_menu_item, goto_hom
 def test_ephemeral_seed_import_tapsigner_real(data, cap_menu, pick_menu_item, goto_home, cap_story,
                                               need_keypress, reset_seed_words, enter_hex, microsd_path,
                                               goto_eph_seed_menu, verify_ephemeral_secret_ui,
-                                              ephemeral_seed_disabled):
+                                              ephemeral_seed_disabled, settings_set):
     fname, backup_key_hex, pub = data
     fpath = microsd_path(fname)
     shutil.copy(f"data/{fname}", fpath)
     reset_seed_words()
+    settings_set("seedvault", None)
     try:
         goto_eph_seed_menu()
     except:
@@ -384,13 +470,20 @@ def test_ephemeral_seed_import_tapsigner_real(data, cap_menu, pick_menu_item, go
 
 
 @pytest.mark.parametrize("way", ["sd", "vdisk", "nfc"])
-@pytest.mark.parametrize('retry', range(3))
+@pytest.mark.parametrize('retry', range(2))
 @pytest.mark.parametrize("testnet", [True, False])
+@pytest.mark.parametrize("seed_vault", [False, True])
 def test_ephemeral_seed_import_xprv(way, retry, testnet, cap_menu, pick_menu_item, goto_home,
                                     cap_story, need_keypress, reset_seed_words, goto_eph_seed_menu,
                                     nfc_write_text, enter_hex, microsd_path, virtdisk_path,
-                                    verify_ephemeral_secret_ui, ephemeral_seed_disabled):
+                                    verify_ephemeral_secret_ui, ephemeral_seed_disabled,
+                                    seed_vault, settings_set):
     reset_seed_words()
+    if seed_vault:
+        settings_set("seedvault", True)
+    else:
+        settings_set("seedvault", None)
+
     fname = "ek.txt"
     from pycoin.key.BIP32Node import BIP32Node
     node = BIP32Node.from_master_secret(os.urandom(32), netcode="XTN" if testnet else "BTC")
@@ -442,6 +535,107 @@ def test_ephemeral_seed_import_xprv(way, retry, testnet, cap_menu, pick_menu_ite
         need_keypress("y")
         pick_menu_item(fname)
 
-    verify_ephemeral_secret_ui(xpub=node.hwif())
+    if seed_vault:
+        time.sleep(.1)
+        _, story = cap_story()
+        assert "Press (1) to store ephemeral secret into Seed Vault" in story
+        need_keypress("1")  # store it
+        need_keypress("y")  # confirm saved to Seed Vault
+
+    verify_ephemeral_secret_ui(xpub=node.hwif(), seed_vault=seed_vault)
+
+
+@pytest.mark.parametrize('data', [
+    [("47649253", "344f9dc08e88b8a46d4b8f46c4e6bb6c"), "crowd language ice brown merit fall release impose egg cheese put suit"],
+    [("CC7BB706", "88f53ed897cc371ffe4b715c267206f3286ed2f655ba9d68"), "material prepare renew convince sell morning weird hotel found crime like town manage harvest sun resemble output dolphin"],
+    [("AC39935C", "956f484cc2136178fd1ad45faeb54972c829f65aad0d74eb2541b11984655893"), "nice kid basket loud current round virtual fold garden interest false tortoise little will height payment insane float expire giraffe obscure crawl girl glare"]
+])
+def test_seed_vault(dev, data, settings_set, settings_get, pick_menu_item, need_keypress, cap_story,
+                    cap_menu, reset_seed_words, get_identity_story, get_seed_value_ux, fake_txn,
+                    try_sign, sim_exec, goto_home, goto_eph_seed_menu):
+    cache_data, mnemonic = data
+    xfp, entropy = cache_data
+    entropy_bytes = bytes.fromhex(entropy)
+    vlen = len(entropy_bytes)
+    assert vlen in [16, 24, 32]
+    marker = 0x80 | ((vlen // 8) - 2)
+    stored_secret = bytes([marker]) + entropy_bytes
+    settings_set("seedvault", None)
+    settings_set("seeds", [(xfp, stored_secret.hex(), f"[{xfp}]")])
+    # enable Seed Vault
+    goto_home()
+    pick_menu_item("Advanced/Tools")
+    pick_menu_item("Danger Zone")
+    pick_menu_item("Seed Vault")
+    time.sleep(.1)
+    _, story = cap_story()
+    assert "Enable Seed Vault?" in story
+    need_keypress("y")
+    time.sleep(.1)
+    pick_menu_item("Enable")
+    time.sleep(.5)
+    goto_home()
+    pick_menu_item("Seed Vault")
+    time.sleep(.1)
+    m = cap_menu()
+    assert len(m) == 1
+    assert xfp in m[0]
+    pick_menu_item(m[0])
+    # rename
+    pick_menu_item("Rename")
+    for _ in range(len(xfp) + 1):  # [xfp]
+        need_keypress("x")
+
+    # below should yield AAAA
+    need_keypress("1")
+    for _ in range(3):
+        need_keypress("9")  # next char
+        need_keypress("1")  # letters
+
+    need_keypress("y")
+    m = cap_menu()
+    assert m[0] == "AAAA"
+    need_keypress("x")  # go back
+    m = cap_menu()
+    assert "AAAA" in m[0]
+    pick_menu_item(m[0])
+    pick_menu_item("Use This Seed")
+    time.sleep(.1)
+    _, story = cap_story()
+    assert xfp in story
+    assert "key in effect until next power down." in story
+    need_keypress("y")
+    active_mnemonic = get_seed_value_ux()
+    assert active_mnemonic == mnemonic.split()
+    istory = get_identity_story()
+    assert "Ephemeral seed is in effect" in istory
+
+    ident_xfp = istory.split("\n\n")[1].strip()
+    assert ident_xfp == xfp
+
+    e_master_xpub = dev.send_recv(CCProtocolPacker.get_xpub(), timeout=5000)
+    assert e_master_xpub != simulator_fixed_xpub
+    psbt = fake_txn(2, 2, master_xpub=e_master_xpub, segwit_in=True)
+    try_sign(psbt, accept=True, finalize=True)  # MUST NOT raise
+    need_keypress("y")
+
+    encoded = sim_exec('from pincodes import pa; RV.write(repr(pa.fetch()))')
+    assert 'Error' not in encoded
+    encoded = eval(encoded)
+    assert len(encoded) == 72
+    assert encoded[0:len(stored_secret)] == stored_secret
+
+    # check rename worked
+    seeds = settings_get("seeds")
+    assert len(seeds) == 1
+    entry = seeds[0]
+    assert len(entry) == 3
+    assert entry[0] == xfp
+    assert entry[1] == stored_secret.hex()
+    assert entry[2] == "AAAA"
+
+    reset_seed_words()
+    time.sleep(.2)
+    need_keypress("x")
 
 # EOF